+++ src/features/communications_func.py
@@ -1,4 +1,4 @@
-from src.context import CallContext
+from src.call_context import CallContext
 from telebot.apihelper import ApiTelegramException
 from src.common_modules.markups import back_transition, markup_transitions
 from src.common_modules.custom_sender import try_to_send
